Project/requirements.txt
appnope==0.1.2
asgiref==3.4.1
backcall==0.2.0
decorator==5.1.0
Django==3.2.6
django-cities==0.6
django-countries==7.2.1
django-model-utils==4.2.0
django-utils-six==2.0
djangorestframework==3.12.4
djangorestframework-simplejwt==5.0.0
ipdb==0.13.9
ipython==7.28.0
jedi==0.18.0
matplotlib-inline==0.1.3
parso==0.8.2
passwords==0.2.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.3.1
prompt-toolkit==3.0.20
psycopg2-binary==2.9.1
ptyprocess==0.7.0
Pygments==2.10.0
PyJWT==2.3.0
pytz==2021.1
six==1.16.0
sqlparse==0.4.1
swapper==1.2.0
toml==0.10.2
tqdm==4.62.3
traitlets==5.1.0
wcwidth==0.2.5
